Create problem3
This program generates a work table of any url listing a keyword and how many times it appears on the page. It can take is multiple urls at the same time. A word only contains the character A-Z, a-z, or 0-9. The files will be stored as url1.txt, url2.txt and etc.